Proof of consept: Implement gateway for SMS bootstrapping
Gateway extracts CoAP package, assuming SMS received is coded with WSP.

+++ include/modem/sms.h
@@ -209,7 +209,7 @@ void sms_unregister_listener(int handle);
  */
 int sms_send_text(const char *number, const char *text);
 
-int sms_send_data(const char *number, const uint8_t *data, uint8_t data_len, const char *udh_str);
+int sms_send_data(const char *number, const uint8_t *data, uint16_t data_len, const char *udh_str);
 
 /** @} */
